[DDW-821] Consolidates Status helpers
