+++ tests/Status/steps/cardano-steps.js
@@ -1,10 +1,7 @@
 // @flow
 import { Given, Then } from 'cucumber';
-import { CardanoNodeStates } from '../../../../source/common/types/cardano-node.types';
-import {
-  getCardanoNodeState,
-  waitForCardanoNodeToExit,
-} from '../helpers/cardano-node-helpers';
+import { CardanoNodeStates } from '../../../source/common/types/cardano-node.types';
+import { getCardanoNodeState, waitForCardanoNodeToExit } from '../helpers';
 
 Given(/^cardano-node is running$/, async function() {
   await this.client.waitUntil(
